in_node_exporter: more pluggable structure
