+++ plugins/in_node_exporter_metrics/CMakeLists.txt
@@ -4,25 +4,19 @@ set(src
   ne_diskstats.c
   ne_filesystem.c
   ne_uname.c
-  ne_stat_linux.c
-  ne_vmstat_linux.c
+  ne_stat.c
+  ne_vmstat.c
   ne_netdev.c
   ne_time.c
   ne_loadavg.c
-  ne_filefd_linux.c
+  ne_filefd.c
   ne_textfile.c
   ne_processes.c
   ne_utils.c
   ne_config.c
+  ne_systemd.c
   ne.c
   )
-
-if(FLB_HAVE_SYSTEMD_SDBUS)
-set(src
-    ${src}
-    ne_systemd.c
-   )
-endif()
 
 if (APPLE)
   FLB_PLUGIN(in_node_exporter_metrics "${src}" "-framework Foundation -framework IOKit")
