+++ plugins/in_node_exporter_metrics/ne.c
@@ -25,7 +25,7 @@
 
 #include "ne.h"
 #include "ne_config.h"
-#include "ne_filefd_linux.h"
+#include "ne_filefd.h"
 
 /* collectors */
 #include "ne_cpu.h"
@@ -34,191 +34,14 @@
 #include "ne_diskstats.h"
 #include "ne_filesystem.h"
 #include "ne_uname.h"
-#include "ne_stat_linux.h"
+#include "ne_stat.h"
 #include "ne_time.h"
 #include "ne_loadavg.h"
-#include "ne_vmstat_linux.h"
+#include "ne_vmstat.h"
 #include "ne_netdev.h"
 #include "ne_textfile.h"
 #include "ne_systemd.h"
 #include "ne_processes.h"
-
-static int ne_timer_cpu_metrics_cb(struct flb_input_instance *ins,
-                                   struct flb_config *config, void *in_context)
-{
-    struct flb_ne *ctx = in_context;
-
-    ne_cpu_update(ctx);
-
-    return 0;
-}
-
-static int ne_timer_cpufreq_metrics_cb(struct flb_input_instance *ins,
-                                       struct flb_config *config, void *in_context)
-{
-    struct flb_ne *ctx = in_context;
-
-    ne_cpufreq_update(ctx);
-
-    return 0;
-}
-
-static int ne_timer_meminfo_metrics_cb(struct flb_input_instance *ins,
-                                       struct flb_config *config, void *in_context)
-{
-    struct flb_ne *ctx = in_context;
-
-    ne_meminfo_update(ctx);
-
-    return 0;
-}
-
-static int ne_timer_diskstats_metrics_cb(struct flb_input_instance *ins,
-                                         struct flb_config *config, void *in_context)
-{
-    struct flb_ne *ctx = in_context;
-
-    ne_diskstats_update(ctx);
-
-    return 0;
-}
-
-static int ne_timer_filesystem_metrics_cb(struct flb_input_instance *ins,
-                                          struct flb_config *config, void *in_context)
-{
-    struct flb_ne *ctx = in_context;
-
-    ne_filesystem_update(ctx);
-
-    return 0;
-}
-
-static int ne_timer_uname_metrics_cb(struct flb_input_instance *ins,
-                                     struct flb_config *config, void *in_context)
-{
-    struct flb_ne *ctx = in_context;
-
-    ne_uname_update(ctx);
-
-    return 0;
-}
-
-static int ne_timer_stat_metrics_cb(struct flb_input_instance *ins,
-                                    struct flb_config *config, void *in_context)
-{
-    struct flb_ne *ctx = in_context;
-
-    ne_stat_update(ctx);
-
-    return 0;
-}
-
-static int ne_timer_time_metrics_cb(struct flb_input_instance *ins,
-                                    struct flb_config *config, void *in_context)
-{
-    struct flb_ne *ctx = in_context;
-
-    ne_time_update(ctx);
-
-    return 0;
-}
-
-static int ne_timer_loadavg_metrics_cb(struct flb_input_instance *ins,
-                                       struct flb_config *config, void *in_context)
-{
-    struct flb_ne *ctx = in_context;
-
-    ne_loadavg_update(ctx);
-
-    return 0;
-}
-
-static int ne_timer_vmstat_metrics_cb(struct flb_input_instance *ins,
-                                      struct flb_config *config, void *in_context)
-{
-    struct flb_ne *ctx = in_context;
-
-    ne_vmstat_update(ctx);
-
-    return 0;
-}
-
-static int ne_timer_netdev_metrics_cb(struct flb_input_instance *ins,
-                                      struct flb_config *config, void *in_context)
-{
-    struct flb_ne *ctx = in_context;
-
-    ne_netdev_update(ctx);
-
-    return 0;
-}
-
-static int ne_timer_filefd_metrics_cb(struct flb_input_instance *ins,
-                                      struct flb_config *config, void *in_context)
-{
-    struct flb_ne *ctx = in_context;
-
-    ne_filefd_update(ctx);
-
-    return 0;
-}
-
-static int ne_timer_textfile_metrics_cb(struct flb_input_instance *ins,
-                                        struct flb_config *config, void *in_context)
-{
-    struct flb_ne *ctx = in_context;
-
-    ne_textfile_update(ctx);
-
-    return 0;
-}
-
-static int ne_timer_systemd_metrics_cb(struct flb_input_instance *ins,
-                                       struct flb_config *config, void *in_context)
-{
-    struct flb_ne *ctx = in_context;
-
-    ne_systemd_update(ctx);
-
-    return 0;
-}
-
-static int ne_timer_processes_metrics_cb(struct flb_input_instance *ins,
-                                       struct flb_config *config, void *in_context)
-{
-    struct flb_ne *ctx = in_context;
-
-    ne_processes_update(ctx);
-
-    return 0;
-}
-struct flb_ne_callback {
-    char *name;
-    void (*func)(char *, void *, void *);
-};
-
-static int ne_update_cb(struct flb_ne *ctx, char *name);
-
-static void update_metrics(struct flb_input_instance *ins, struct flb_ne *ctx)
-{
-    int ret;
-    struct mk_list *head;
-    struct flb_slist_entry *entry;
-
-    /* Update our metrics */
-    if (ctx->metrics) {
-        mk_list_foreach(head, ctx->metrics) {
-            entry = mk_list_entry(head, struct flb_slist_entry, _head);
-            ret = flb_callback_exists(ctx->callback, entry->str);
-            if (ret == FLB_TRUE) {
-                ne_update_cb(ctx, entry->str);
-            }
-            else {
-                flb_plg_debug(ctx->ins, "Callback for metrics '%s' is not registered", entry->str);
-            }
-        }
-    }
-}
 
 /*
  * Update the metrics, this function is invoked every time 'scrape_interval'
@@ -230,8 +53,6 @@ static int cb_ne_collect(struct flb_input_instance *ins,
     int ret;
     struct flb_ne *ctx = in_context;
 
-    update_metrics(ins, ctx);
-
     /* Append the updated metrics */
     ret = flb_input_metrics_append(ins, NULL, 0, ctx->cmt);
     if (ret != 0) {
@@ -241,151 +62,110 @@ static int cb_ne_collect(struct flb_input_instance *ins,
     return 0;
 }
 
-static void ne_cpu_update_cb(char *name, void *p1, void *p2)
+static int collectors_common_init(struct flb_ne_collector *coll)
 {
-    struct flb_ne *ctx = p1;
+    if (coll == NULL) {
+        return -1;
+    }
+    coll->coll_fd = -1;
+    coll->interval = 0;
+    coll->activated = FLB_FALSE;
 
-    ne_cpu_update(ctx);
+    return 0;
 }
 
-static void ne_cpufreq_update_cb(char *name, void *p1, void *p2)
+static int get_interval_property(struct flb_ne *ctx, flb_sds_t name)
 {
-    struct flb_ne *ctx = p1;
+    flb_sds_t interval_conf_name;
+    size_t conf_len = 1024;
+    const char *interval_str;
+    int ret;
+    int interval;
 
-    ne_cpufreq_update(ctx);
+    interval_conf_name = flb_sds_create_size(conf_len);
+    if (interval_conf_name == NULL) {
+        flb_errno();
+        return -1;
+    }
+    ret = flb_sds_snprintf(&interval_conf_name, conf_len, "collector.%s.scrape_interval", name);
+    if (ret < 0) {
+        flb_errno();
+        flb_sds_destroy(interval_conf_name);
+        return -1;
+    }
+    else if (ret > conf_len) {
+        flb_plg_error(ctx->ins, "buffer is small for %s interval config", name);
+        flb_sds_destroy(interval_conf_name);
+        return -1;
+    }
+
+    interval_str = flb_input_get_property(interval_conf_name, ctx->ins);
+    if (interval_str == NULL) {
+        interval = ctx->scrape_interval;
+    }
+    else {
+        interval = atoi(interval_str);
+        if (interval == 0) {
+            interval = ctx->scrape_interval;
+        }
+    }
+    flb_sds_destroy(interval_conf_name);
+
+    return interval;
 }
 
-static void ne_meminfo_update_cb(char *name, void *p1, void *p2)
+static int activate_collector(struct flb_ne *ctx, struct flb_config *config, 
+                              struct flb_ne_collector *coll, flb_sds_t name)
 {
-    struct flb_ne *ctx = p1;
-
-    ne_meminfo_update(ctx);
-}
-
-static void ne_diskstats_update_cb(char *name, void *p1, void *p2)
-{
-    struct flb_ne *ctx = p1;
-
-    ne_diskstats_update(ctx);
-}
-
-static void ne_filesystem_update_cb(char *name, void *p1, void *p2)
-{
-    struct flb_ne *ctx = p1;
-
-    ne_filesystem_update(ctx);
-}
-
-static void ne_uname_update_cb(char *name, void *p1, void *p2)
-{
-    struct flb_ne *ctx = p1;
-
-    ne_uname_update(ctx);
-}
-
-static void ne_stat_update_cb(char *name, void *p1, void *p2)
-{
-    struct flb_ne *ctx = p1;
-
-    ne_stat_update(ctx);
-}
-
-static void ne_time_update_cb(char *name, void *p1, void *p2)
-{
-    struct flb_ne *ctx = p1;
-
-    ne_time_update(ctx);
-}
-
-static void ne_loadavg_update_cb(char *name, void *p1, void *p2)
-{
-    struct flb_ne *ctx = p1;
-
-    ne_loadavg_update(ctx);
-}
-
-static void ne_vmstat_update_cb(char *name, void *p1, void *p2)
-{
-    struct flb_ne *ctx = p1;
-
-    ne_vmstat_update(ctx);
-}
-
-static void ne_netdev_update_cb(char *name, void *p1, void *p2)
-{
-    struct flb_ne *ctx = p1;
-
-    ne_netdev_update(ctx);
-}
-
-static void ne_filefd_update_cb(char *name, void *p1, void *p2)
-{
-    struct flb_ne *ctx = p1;
-
-    ne_filefd_update(ctx);
-}
-
-static void ne_textfile_update_cb(char *name, void *p1, void *p2)
-{
-    struct flb_ne *ctx = p1;
-
-    ne_textfile_update(ctx);
-}
-
-static void ne_systemd_update_cb(char *name, void *p1, void *p2)
-{
-    struct flb_ne *ctx = p1;
-
-    ne_systemd_update(ctx);
-}
-
-static void ne_processes_update_cb(char *name, void *p1, void *p2)
-{
-    struct flb_ne *ctx = p1;
-
-    ne_processes_update(ctx);
-}
-
-static int ne_update_cb(struct flb_ne *ctx, char *name)
-{
+    int interval;
     int ret;
 
-    ret = flb_callback_do(ctx->callback, name, ctx, NULL);
-    return ret;
-}
+    if (coll == NULL) {
+        return -1;
+    }
+    if (coll->activated == FLB_TRUE) {
+        flb_plg_warn(ctx->ins, "%s is already activated", name);
+        return 0;
+    }
+    if (coll->cb_init == NULL) {
+        flb_plg_warn(ctx->ins, "%s is not supported", name);
+        return 0;
+    }
 
-/*
- * Callbacks Table
- */
-struct flb_ne_callback ne_callbacks[] = {
-    /* metrics */
-    { "cpufreq", ne_cpufreq_update_cb },
-    { "cpu", ne_cpu_update_cb },
-    { "meminfo", ne_meminfo_update_cb },
-    { "diskstats", ne_diskstats_update_cb },
-    { "filesystem", ne_filesystem_update_cb },
-    { "uname", ne_uname_update_cb },
-    { "stat", ne_stat_update_cb },
-    { "time", ne_time_update_cb },
-    { "loadavg", ne_loadavg_update_cb },
-    { "vmstat", ne_vmstat_update_cb },
-    { "netdev", ne_netdev_update_cb },
-    { "filefd", ne_filefd_update_cb },
-    { "textfile", ne_textfile_update_cb },
-    { "systemd", ne_systemd_update_cb },
-    { "processes", ne_processes_update_cb },
-    { 0 }
-};
+    if (coll->cb_update) {
+        interval = get_interval_property(ctx, name);
+        if (interval < 0) {
+            return -1;
+        }
+        ret = flb_input_set_collector_time(ctx->ins,
+                                           coll->cb_update,
+                                           interval, 0, config);
+        if (ret < 0) {
+            flb_plg_error(ctx->ins, "flb_input_set_collector_time failed");
+            return -1;
+        }
+        coll->coll_fd = ret;
+    }
+
+    ret = coll->cb_init(ctx);
+    if (ret != 0) {
+        flb_plg_error(ctx->ins, "%s init failed", name);
+        return -1;
+    }
+    coll->activated = FLB_TRUE;
+
+    return 0;
+}
 
 static int in_ne_init(struct flb_input_instance *in,
                       struct flb_config *config, void *data)
 {
     int ret;
-    int metric_idx = -1;
     struct flb_ne *ctx;
     struct mk_list *head;
+    struct mk_list *coll_head;
+    struct flb_ne_collector *coll;
     struct flb_slist_entry *entry;
-    struct flb_ne_callback *cb;
 
     /* Create plugin context */
     ctx = flb_ne_config_create(in, config);
@@ -394,29 +174,30 @@ static int in_ne_init(struct flb_input_instance *in,
         return -1;
     }
 
+    mk_list_init(&ctx->collectors);
+    mk_list_add(&cpu_collector._head, &ctx->collectors);
+    mk_list_add(&cpufreq_collector._head, &ctx->collectors);
+    mk_list_add(&meminfo_collector._head, &ctx->collectors);
+    mk_list_add(&diskstats_collector._head, &ctx->collectors);
+    mk_list_add(&filesystem_collector._head, &ctx->collectors);
+    mk_list_add(&uname_collector._head, &ctx->collectors);
+    mk_list_add(&stat_collector._head, &ctx->collectors);
+    mk_list_add(&time_collector._head, &ctx->collectors);
+    mk_list_add(&loadavg_collector._head, &ctx->collectors);
+    mk_list_add(&vmstat_collector._head, &ctx->collectors);
+    mk_list_add(&netdev_collector._head, &ctx->collectors);
+    mk_list_add(&filefd_collector._head, &ctx->collectors);
+    mk_list_add(&textfile_collector._head, &ctx->collectors);
+    mk_list_add(&systemd_collector._head, &ctx->collectors);
+    mk_list_add(&processes_collector._head, &ctx->collectors);
+
+    mk_list_foreach(head, &ctx->collectors) {
+        coll = mk_list_entry(head, struct flb_ne_collector, _head);
+        collectors_common_init(coll);
+    }
+
     /* Initialize fds */
     ctx->coll_fd = -1;
-    ctx->coll_cpu_fd = -1;
-    ctx->coll_cpufreq_fd = -1;
-    ctx->coll_meminfo_fd = -1;
-    ctx->coll_diskstats_fd = -1;
-    ctx->coll_filesystem_fd = -1;
-    ctx->coll_uname_fd = -1;
-    ctx->coll_stat_fd = -1;
-    ctx->coll_time_fd = -1;
-    ctx->coll_loadavg_fd = -1;
-    ctx->coll_vmstat_fd = -1;
-    ctx->coll_netdev_fd = -1;
-    ctx->coll_filefd_fd = -1;
-    ctx->coll_textfile_fd = -1;
-    ctx->coll_systemd_fd = -1;
-    ctx->coll_processes_fd = -1;
-
-    ctx->callback = flb_callback_create(in->name);
-    if (!ctx->callback) {
-        flb_plg_error(ctx->ins, "Create callback failed");
-        return -1;
-    }
 
     /* Associate context with the instance */
     flb_input_set_context(in, ctx);
@@ -437,321 +218,21 @@ static int in_ne_init(struct flb_input_instance *in,
     if (ctx->metrics) {
         mk_list_foreach(head, ctx->metrics) {
             entry = mk_list_entry(head, struct flb_slist_entry, _head);
-            ret = flb_callback_exists(ctx->callback, entry->str);
 
-            if (ret == FLB_FALSE) {
-                if (strncmp(entry->str, "cpufreq", 7) == 0) {
-                    if (ctx->cpufreq_scrape_interval == 0) {
+            mk_list_foreach(coll_head, &ctx->collectors) {
+                coll = mk_list_entry(coll_head, struct flb_ne_collector, _head);
+                if (coll->activated == FLB_FALSE &&
+                    flb_sds_len(entry->str) == strlen(coll->name) &&
+                    strncmp(entry->str, coll->name, strlen(coll->name)) == 0) {
+                    ret = activate_collector(ctx, config, coll, entry->str);
+                    if (ret < 0) {
+                        flb_plg_error(ctx->ins,
+                                      "could not set %s collector for Node Exporter Metrics plugin", entry->str);
+                    }
+                    else {
                         flb_plg_debug(ctx->ins, "enabled metrics %s", entry->str);
-                        metric_idx = 0;
                     }
-                    else if (ctx->cpufreq_scrape_interval > 0) {
-                        /* Create the cpufreq collector */
-                        ret = flb_input_set_collector_time(in,
-                                                           ne_timer_cpufreq_metrics_cb,
-                                                           ctx->cpufreq_scrape_interval, 0,
-                                                           config);
-                        if (ret == -1) {
-                            flb_plg_error(ctx->ins,
-                                          "could not set cpufreq collector for Node Exporter Metrics plugin");
-                            return -1;
-                        }
-                        ctx->coll_cpufreq_fd = ret;
-                    }
-                    ne_cpufreq_init(ctx);
-                }
-                else if (strncmp(entry->str, "cpu", 3) == 0) {
-                    if (ctx->cpu_scrape_interval == 0) {
-                        flb_plg_debug(ctx->ins, "enabled metrics %s", entry->str);
-                        metric_idx = 1;
-                    }
-                    else if (ctx->cpu_scrape_interval > 0) {
-                        /* Create the cpu collector */
-                        ret = flb_input_set_collector_time(in,
-                                                           ne_timer_cpu_metrics_cb,
-                                                           ctx->cpu_scrape_interval, 0,
-                                                           config);
-                        if (ret == -1) {
-                            flb_plg_error(ctx->ins,
-                                          "could not set cpu collector for Node Exporter Metrics plugin");
-                            return -1;
-                        }
-                        ctx->coll_cpu_fd = ret;
-                    }
-                    ne_cpu_init(ctx);
-                }
-                else if (strncmp(entry->str, "meminfo", 7) == 0) {
-                    if (ctx->meminfo_scrape_interval == 0) {
-                        flb_plg_debug(ctx->ins, "enabled metrics %s", entry->str);
-                        metric_idx = 2;
-                    }
-                    else if (ctx->meminfo_scrape_interval > 0) {
-                        /* Create the meminfo collector */
-                        ret = flb_input_set_collector_time(in,
-                                                           ne_timer_meminfo_metrics_cb,
-                                                           ctx->meminfo_scrape_interval, 0,
-                                                           config);
-                        if (ret == -1) {
-                            flb_plg_error(ctx->ins,
-                                          "could not set meminfo collector for Node Exporter Metrics plugin");
-                            return -1;
-                        }
-                        ctx->coll_meminfo_fd = ret;
-                    }
-                    ne_meminfo_init(ctx);
-                }
-                else if (strncmp(entry->str, "diskstats", 9) == 0) {
-                    if (ctx->diskstats_scrape_interval == 0) {
-                        flb_plg_debug(ctx->ins, "enabled metrics %s", entry->str);
-                        metric_idx = 3;
-                    }
-                    else if (ctx->diskstats_scrape_interval > 0) {
-                        /* Create the diskstats collector */
-                        ret = flb_input_set_collector_time(in,
-                                                           ne_timer_diskstats_metrics_cb,
-                                                           ctx->diskstats_scrape_interval, 0,
-                                                           config);
-                        if (ret == -1) {
-                            flb_plg_error(ctx->ins,
-                                          "could not set diskstats collector for Node Exporter Metrics plugin");
-                            return -1;
-                        }
-                        ctx->coll_diskstats_fd = ret;
-                    }
-                    ne_diskstats_init(ctx);
-                }
-                else if (strncmp(entry->str, "filesystem", 10) == 0) {
-                    if (ctx->diskstats_scrape_interval == 0) {
-                        flb_plg_debug(ctx->ins, "enabled metrics %s", entry->str);
-                        metric_idx = 4;
-                    }
-                    else if (ctx->filesystem_scrape_interval > 0) {
-                        /* Create the diskstats collector */
-                        ret = flb_input_set_collector_time(in,
-                                                           ne_timer_filesystem_metrics_cb,
-                                                           ctx->filesystem_scrape_interval, 0,
-                                                           config);
-                        if (ret == -1) {
-                            flb_plg_error(ctx->ins,
-                                          "could not set filesystem collector for Node Exporter Metrics plugin");
-                            return -1;
-                        }
-                        ctx->coll_filesystem_fd = ret;
-                    }
-                    ne_filesystem_init(ctx);
-                }
-                else if (strncmp(entry->str, "uname", 5) == 0) {
-                    if (ctx->uname_scrape_interval == 0) {
-                        flb_plg_debug(ctx->ins, "enabled metrics %s", entry->str);
-                        metric_idx = 5;
-                    }
-                    else if (ctx->uname_scrape_interval > 0) {
-                        /* Create the uname collector */
-                        ret = flb_input_set_collector_time(in,
-                                                           ne_timer_uname_metrics_cb,
-                                                           ctx->uname_scrape_interval, 0,
-                                                           config);
-                        if (ret == -1) {
-                            flb_plg_error(ctx->ins,
-                                          "could not set uname collector for Node Exporter Metrics plugin");
-                            return -1;
-                        }
-                        ctx->coll_uname_fd = ret;
-                    }
-                    ne_uname_init(ctx);
-                }
-                else if (strncmp(entry->str, "stat", 4) == 0) {
-                    if (ctx->stat_scrape_interval == 0) {
-                        flb_plg_debug(ctx->ins, "enabled metrics %s", entry->str);
-                        metric_idx = 6;
-                    }
-                    else if (ctx->stat_scrape_interval > 0) {
-                        /* Create the meminfo collector */
-                        ret = flb_input_set_collector_time(in,
-                                                           ne_timer_stat_metrics_cb,
-                                                           ctx->stat_scrape_interval, 0,
-                                                           config);
-                        if (ret == -1) {
-                            flb_plg_error(ctx->ins,
-                                          "could not set meminfo collector for Node Exporter Metrics plugin");
-                            return -1;
-                        }
-                        ctx->coll_stat_fd = ret;
-                    }
-                    ne_stat_init(ctx);
-                }
-                else if (strncmp(entry->str, "time", 4) == 0) {
-                    if (ctx->time_scrape_interval == 0) {
-                        flb_plg_debug(ctx->ins, "enabled metrics %s", entry->str);
-                        metric_idx = 7;
-                    }
-                    else if (ctx->time_scrape_interval > 0) {
-                        /* Create the time collector */
-                        ret = flb_input_set_collector_time(in,
-                                                           ne_timer_time_metrics_cb,
-                                                           ctx->time_scrape_interval, 0,
-                                                           config);
-                        if (ret == -1) {
-                            flb_plg_error(ctx->ins,
-                                          "could not set time collector for Node Exporter Metrics plugin");
-                            return -1;
-                        }
-                        ctx->coll_time_fd = ret;
-                    }
-                    ne_time_init(ctx);
-                }
-                else if (strncmp(entry->str, "loadavg", 7) == 0) {
-                    if (ctx->loadavg_scrape_interval == 0) {
-                        flb_plg_debug(ctx->ins, "enabled metrics %s", entry->str);
-                        metric_idx = 8;
-                    }
-                    else if (ctx->loadavg_scrape_interval > 0) {
-                        /* Create the loadavg collector */
-                        ret = flb_input_set_collector_time(in,
-                                                           ne_timer_loadavg_metrics_cb,
-                                                           ctx->loadavg_scrape_interval, 0,
-                                                           config);
-                        if (ret == -1) {
-                            flb_plg_error(ctx->ins,
-                                          "could not set loadavg collector for Node Exporter Metrics plugin");
-                            return -1;
-                        }
-                        ctx->coll_loadavg_fd = ret;
-                    }
-                    ne_loadavg_init(ctx);
-                }
-                else if (strncmp(entry->str, "vmstat", 6) == 0) {
-                    if (ctx->vmstat_scrape_interval == 0) {
-                        flb_plg_debug(ctx->ins, "enabled metrics %s", entry->str);
-                        metric_idx = 9;
-                    }
-                    else if (ctx->vmstat_scrape_interval > 0) {
-                        /* Create the vmstat collector */
-                        ret = flb_input_set_collector_time(in,
-                                                           ne_timer_vmstat_metrics_cb,
-                                                           ctx->vmstat_scrape_interval, 0,
-                                                           config);
-                        if (ret == -1) {
-                            flb_plg_error(ctx->ins,
-                                          "could not set vmstat collector for Node Exporter Metrics plugin");
-                            return -1;
-                        }
-                        ctx->coll_vmstat_fd = ret;
-                    }
-                    ne_vmstat_init(ctx);
-                }
-                else if (strncmp(entry->str, "netdev", 6) == 0) {
-                    if (ctx->netdev_scrape_interval == 0) {
-                        flb_plg_debug(ctx->ins, "enabled metrics %s", entry->str);
-                        metric_idx = 10;
-                    }
-                    else if (ctx->netdev_scrape_interval > 0) {
-                        /* Create the netdev collector */
-                        ret = flb_input_set_collector_time(in,
-                                                           ne_timer_netdev_metrics_cb,
-                                                           ctx->netdev_scrape_interval, 0,
-                                                           config);
-                        if (ret == -1) {
-                            flb_plg_error(ctx->ins,
-                                          "could not set netdev collector for Node Exporter Metrics plugin");
-                            return -1;
-                        }
-                        ctx->coll_netdev_fd = ret;
-                    }
-                    ne_netdev_init(ctx);
-                }
-                else if (strncmp(entry->str, "filefd", 6) == 0) {
-                    if (ctx->filefd_scrape_interval == 0) {
-                        flb_plg_debug(ctx->ins, "enabled metrics %s", entry->str);
-                        metric_idx = 11;
-                    }
-                    else if (ctx->filefd_scrape_interval > 0) {
-                        /* Create the filefd collector */
-                        ret = flb_input_set_collector_time(in,
-                                                           ne_timer_filefd_metrics_cb,
-                                                           ctx->filefd_scrape_interval, 0,
-                                                           config);
-                        if (ret == -1) {
-                            flb_plg_error(ctx->ins,
-                                          "could not set filefd collector for Node Exporter Metrics plugin");
-                            return -1;
-                        }
-                        ctx->coll_filefd_fd = ret;
-                    }
-                    ne_filefd_init(ctx);
-                }
-                else if (strncmp(entry->str, "textfile", 8) == 0) {
-                    if (ctx->textfile_scrape_interval == 0) {
-                        flb_plg_debug(ctx->ins, "enabled metrics %s", entry->str);
-                        metric_idx = 12;
-                    }
-                    else if (ctx->textfile_scrape_interval > 0) {
-                        /* Create the filefd collector */
-                        ret = flb_input_set_collector_time(in,
-                                                           ne_timer_textfile_metrics_cb,
-                                                           ctx->textfile_scrape_interval, 0,
-                                                           config);
-                        if (ret == -1) {
-                            flb_plg_error(ctx->ins,
-                                          "could not set textfile collector for Node Exporter Metrics plugin");
-                            return -1;
-                        }
-                        ctx->coll_textfile_fd = ret;
-                    }
-                    ne_textfile_init(ctx);
-                }
-                else if (strncmp(entry->str, "systemd", 7) == 0) {
-                    if (ctx->systemd_scrape_interval == 0) {
-                        flb_plg_debug(ctx->ins, "enabled metrics %s", entry->str);
-                        metric_idx = 13;
-                    }
-                    else if (ctx->systemd_scrape_interval > 0) {
-                        /* Create the filefd collector */
-                        ret = flb_input_set_collector_time(in,
-                                                           ne_timer_systemd_metrics_cb,
-                                                           ctx->systemd_scrape_interval, 0,
-                                                           config);
-                        if (ret == -1) {
-                            flb_plg_error(ctx->ins,
-                                          "could not set systemd collector for Node Exporter Metrics plugin");
-                            return -1;
-                        }
-                        ctx->coll_systemd_fd = ret;
-                    }
-                    ne_systemd_init(ctx);
-                }
-                else if (strncmp(entry->str, "processes", 9) == 0) {
-                    if (ctx->processes_scrape_interval == 0) {
-                        flb_plg_debug(ctx->ins, "enabled metrics %s", entry->str);
-                        metric_idx = 14;
-                    }
-                    else if (ctx->processes_scrape_interval > 0) {
-                        /* Create the filefd collector */
-                        ret = flb_input_set_collector_time(in,
-                                                           ne_timer_processes_metrics_cb,
-                                                           ctx->processes_scrape_interval, 0,
-                                                           config);
-                        if (ret == -1) {
-                            flb_plg_error(ctx->ins,
-                                          "could not set systemd collector for Node Exporter Metrics plugin");
-                            return -1;
-                        }
-                        ctx->coll_processes_fd = ret;
-                    }
-                    ne_processes_init(ctx);
-                }
-                else {
-                    flb_plg_warn(ctx->ins, "Unknown metrics: %s", entry->str);
-                    metric_idx = -1;
-                }
-
-                if (metric_idx >= 0) {
-                    cb = &ne_callbacks[metric_idx];
-                    ret = flb_callback_set(ctx->callback, cb->name, cb->func);
-                    if (ret == -1) {
-                        flb_plg_error(ctx->ins, "error setting up default "
-                                      "callback '%s'", cb->name);
-                    }
+                    break;
                 }
             }
         }
@@ -767,101 +248,22 @@ static int in_ne_init(struct flb_input_instance *in,
 
 static int in_ne_exit(void *data, struct flb_config *config)
 {
-    int ret;
     struct flb_ne *ctx = data;
-    struct mk_list *head;
-    struct flb_slist_entry *entry;
+    struct mk_list *coll_head;
+    struct flb_ne_collector *coll;
 
     if (!ctx) {
         return 0;
     }
 
-    /* Teardown for callback tied up resources */
-    if (ctx->metrics) {
-        mk_list_foreach(head, ctx->metrics) {
-            entry = mk_list_entry(head, struct flb_slist_entry, _head);
-            ret = flb_callback_exists(ctx->callback, entry->str);
-
-            if (ret == FLB_TRUE) {
-                if (strncmp(entry->str, "cpufreq", 7) == 0) {
-                    /* nop */
-                }
-                else if (strncmp(entry->str, "cpu", 3) == 0) {
-                    /* nop */
-                }
-                else if (strncmp(entry->str, "meminfo", 7) == 0) {
-                    ne_meminfo_exit(ctx);
-                }
-                else if (strncmp(entry->str, "diskstats", 9) == 0) {
-                    ne_diskstats_exit(ctx);
-                }
-                else if (strncmp(entry->str, "filesystem", 10) == 0) {
-                    ne_filesystem_exit(ctx);
-                }
-                else if (strncmp(entry->str, "uname", 5) == 0) {
-                    /* nop */
-                }
-                else if (strncmp(entry->str, "stat", 4) == 0) {
-                    /* nop */
-                }
-                else if (strncmp(entry->str, "time", 4) == 0) {
-                    /* nop */
-                }
-                else if (strncmp(entry->str, "loadavg", 7) == 0) {
-                    /* nop */
-                }
-                else if (strncmp(entry->str, "vmstat", 6) == 0) {
-                    ne_vmstat_exit(ctx);
-                }
-                else if (strncmp(entry->str, "netdev", 6) == 0) {
-                    ne_netdev_exit(ctx);
-                }
-                else if (strncmp(entry->str, "filefd", 6) == 0) {
-                    /* nop */
-                }
-                else if (strncmp(entry->str, "textfile", 8) == 0) {
-                    /* nop */
-                }
-                else if (strncmp(entry->str, "systemd", 7) == 0) {
-                    ne_systemd_exit(ctx);
-                }
-                else if (strncmp(entry->str, "processes", 9) == 0) {
-                    ne_processes_exit(ctx);
-                }
-                else {
-                    flb_plg_warn(ctx->ins, "Unknown metrics: %s", entry->str);
-                }
-            }
+    mk_list_foreach(coll_head, &ctx->collectors) {
+        coll = mk_list_entry(coll_head, struct flb_ne_collector, _head);
+        if (coll->activated == FLB_TRUE && coll->cb_exit) {
+            coll->cb_exit(ctx);
         }
     }
 
-    /* destroy callback context */
-    if (ctx->callback) {
-        flb_callback_destroy(ctx->callback);
-    }
-
     /* Teardown for timer tied up resources */
-    if (ctx->coll_meminfo_fd != -1) {
-        ne_meminfo_exit(ctx);
-    }
-    if (ctx->coll_diskstats_fd != -1) {
-        ne_diskstats_exit(ctx);
-    }
-    if (ctx->coll_filesystem_fd != -1) {
-        ne_filesystem_exit(ctx);
-    }
-    if (ctx->coll_vmstat_fd != -1) {
-        ne_vmstat_exit(ctx);
-    }
-    if (ctx->coll_netdev_fd != -1) {
-        ne_netdev_exit(ctx);
-    }
-    if (ctx->coll_systemd_fd != -1) {
-        ne_systemd_exit(ctx);
-    }
-    if (ctx->coll_processes_fd != -1) {
-        ne_processes_exit(ctx);
-    }
 
     flb_ne_config_destroy(ctx);
 
@@ -871,104 +273,33 @@ static int in_ne_exit(void *data, struct flb_config *config)
 static void in_ne_pause(void *data, struct flb_config *config)
 {
     struct flb_ne *ctx = data;
+    struct flb_ne_collector *coll;
+    struct mk_list *head;
 
     flb_input_collector_pause(ctx->coll_fd, ctx->ins);
-    if (ctx->coll_cpu_fd != -1) {
-        flb_input_collector_pause(ctx->coll_cpu_fd, ctx->ins);
-    }
-    if (ctx->coll_cpufreq_fd != -1) {
-        flb_input_collector_pause(ctx->coll_cpufreq_fd, ctx->ins);
-    }
-    if (ctx->coll_meminfo_fd != -1) {
-        flb_input_collector_pause(ctx->coll_meminfo_fd, ctx->ins);
-    }
-    if (ctx->coll_diskstats_fd != -1) {
-        flb_input_collector_pause(ctx->coll_diskstats_fd, ctx->ins);
-    }
-    if (ctx->coll_filesystem_fd != -1) {
-        flb_input_collector_pause(ctx->coll_filesystem_fd, ctx->ins);
-    }
-    if (ctx->coll_uname_fd != -1) {
-        flb_input_collector_pause(ctx->coll_uname_fd, ctx->ins);
-    }
-    if (ctx->coll_stat_fd != -1) {
-        flb_input_collector_pause(ctx->coll_stat_fd, ctx->ins);
-    }
-    if (ctx->coll_time_fd != -1) {
-        flb_input_collector_pause(ctx->coll_time_fd, ctx->ins);
-    }
-    if (ctx->coll_loadavg_fd != -1) {
-        flb_input_collector_pause(ctx->coll_loadavg_fd, ctx->ins);
-    }
-    if (ctx->coll_vmstat_fd != -1) {
-        flb_input_collector_pause(ctx->coll_vmstat_fd, ctx->ins);
-    }
-    if (ctx->coll_netdev_fd != -1) {
-        flb_input_collector_pause(ctx->coll_netdev_fd, ctx->ins);
-    }
-    if (ctx->coll_filefd_fd != -1) {
-        flb_input_collector_pause(ctx->coll_filefd_fd, ctx->ins);
-    }
-    if (ctx->coll_textfile_fd != -1) {
-        flb_input_collector_pause(ctx->coll_textfile_fd, ctx->ins);
-    }
-    if (ctx->coll_systemd_fd != -1) {
-        flb_input_collector_pause(ctx->coll_systemd_fd, ctx->ins);
-    }
-    if (ctx->coll_processes_fd != -1) {
-        flb_input_collector_pause(ctx->coll_processes_fd, ctx->ins);
+
+    mk_list_foreach(head, &ctx->collectors) {
+        coll = mk_list_entry(head, struct flb_ne_collector, _head);
+        if (coll->activated == FLB_FALSE) {
+            continue;
+        }
+        flb_input_collector_pause(coll->coll_fd, ctx->ins);
     }
 }
 
 static void in_ne_resume(void *data, struct flb_config *config)
 {
     struct flb_ne *ctx = data;
+    struct flb_ne_collector *coll;
+    struct mk_list *head;
 
     flb_input_collector_resume(ctx->coll_fd, ctx->ins);
-    if (ctx->coll_cpu_fd != -1) {
-        flb_input_collector_resume(ctx->coll_cpu_fd, ctx->ins);
-    }
-    if (ctx->coll_cpufreq_fd != -1) {
-        flb_input_collector_resume(ctx->coll_cpufreq_fd, ctx->ins);
-    }
-    if (ctx->coll_meminfo_fd != -1) {
-        flb_input_collector_resume(ctx->coll_meminfo_fd, ctx->ins);
-    }
-    if (ctx->coll_diskstats_fd != -1) {
-        flb_input_collector_resume(ctx->coll_diskstats_fd, ctx->ins);
-    }
-    if (ctx->coll_filesystem_fd != -1) {
-        flb_input_collector_resume(ctx->coll_filesystem_fd, ctx->ins);
-    }
-    if (ctx->coll_uname_fd != -1) {
-        flb_input_collector_resume(ctx->coll_uname_fd, ctx->ins);
-    }
-    if (ctx->coll_stat_fd != -1) {
-        flb_input_collector_resume(ctx->coll_stat_fd, ctx->ins);
-    }
-    if (ctx->coll_time_fd != -1) {
-        flb_input_collector_resume(ctx->coll_time_fd, ctx->ins);
-    }
-    if (ctx->coll_loadavg_fd != -1) {
-        flb_input_collector_resume(ctx->coll_loadavg_fd, ctx->ins);
-    }
-    if (ctx->coll_vmstat_fd != -1) {
-        flb_input_collector_resume(ctx->coll_vmstat_fd, ctx->ins);
-    }
-    if (ctx->coll_netdev_fd != -1) {
-        flb_input_collector_resume(ctx->coll_netdev_fd, ctx->ins);
-    }
-    if (ctx->coll_filefd_fd != -1) {
-        flb_input_collector_resume(ctx->coll_filefd_fd, ctx->ins);
-    }
-    if (ctx->coll_textfile_fd != -1) {
-        flb_input_collector_resume(ctx->coll_textfile_fd, ctx->ins);
-    }
-    if (ctx->coll_systemd_fd != -1) {
-        flb_input_collector_resume(ctx->coll_systemd_fd, ctx->ins);
-    }
-    if (ctx->coll_processes_fd != -1) {
-        flb_input_collector_resume(ctx->coll_processes_fd, ctx->ins);
+    mk_list_foreach(head, &ctx->collectors) {
+        coll = mk_list_entry(head, struct flb_ne_collector, _head);
+        if (coll->activated == FLB_FALSE) {
+            continue;
+        }
+        flb_input_collector_resume(coll->coll_fd, ctx->ins);
     }
 }
 
@@ -982,91 +313,91 @@ static struct flb_config_map config_map[] = {
 
     {
      FLB_CONFIG_MAP_TIME, "collector.cpu.scrape_interval", "0",
-     0, FLB_TRUE, offsetof(struct flb_ne, cpu_scrape_interval),
+     0, FLB_FALSE, 0,
      "scrape interval to collect cpu metrics from the node."
     },
 
     {
      FLB_CONFIG_MAP_TIME, "collector.cpufreq.scrape_interval", "0",
-     0, FLB_TRUE, offsetof(struct flb_ne, cpufreq_scrape_interval),
+     0, FLB_FALSE, 0,
      "scrape interval to collect cpufreq metrics from the node."
     },
 
     {
      FLB_CONFIG_MAP_TIME, "collector.meminfo.scrape_interval", "0",
-     0, FLB_TRUE, offsetof(struct flb_ne, meminfo_scrape_interval),
+     0, FLB_FALSE, 0,
      "scrape interval to collect meminfo metrics from the node."
     },
 
     {
      FLB_CONFIG_MAP_TIME, "collector.diskstats.scrape_interval", "0",
-     0, FLB_TRUE, offsetof(struct flb_ne, diskstats_scrape_interval),
+     0, FLB_FALSE, 0,
      "scrape interval to collect diskstats metrics from the node."
     },
 
     {
      FLB_CONFIG_MAP_TIME, "collector.filesystem.scrape_interval", "0",
-     0, FLB_TRUE, offsetof(struct flb_ne, filesystem_scrape_interval),
+     0, FLB_FALSE, 0,
      "scrape interval to collect filesystem metrics from the node."
     },
 
     {
      FLB_CONFIG_MAP_TIME, "collector.uname.scrape_interval", "0",
-     0, FLB_TRUE, offsetof(struct flb_ne, uname_scrape_interval),
+     0, FLB_FALSE, 0,
      "scrape interval to collect uname metrics from the node."
     },
 
     {
      FLB_CONFIG_MAP_TIME, "collector.stat.scrape_interval", "0",
-     0, FLB_TRUE, offsetof(struct flb_ne, stat_scrape_interval),
+     0, FLB_FALSE, 0,
      "scrape interval to collect stat metrics from the node."
     },
 
     {
      FLB_CONFIG_MAP_TIME, "collector.time.scrape_interval", "0",
-     0, FLB_TRUE, offsetof(struct flb_ne, time_scrape_interval),
+     0, FLB_FALSE, 0,
      "scrape interval to collect time metrics from the node."
     },
 
     {
      FLB_CONFIG_MAP_TIME, "collector.loadavg.scrape_interval", "0",
-     0, FLB_TRUE, offsetof(struct flb_ne, loadavg_scrape_interval),
+     0, FLB_FALSE, 0,
      "scrape interval to collect loadavg metrics from the node."
     },
 
     {
      FLB_CONFIG_MAP_TIME, "collector.vmstat.scrape_interval", "0",
-     0, FLB_TRUE, offsetof(struct flb_ne, vmstat_scrape_interval),
+     0, FLB_FALSE, 0,
      "scrape interval to collect vmstat metrics from the node."
     },
 
     {
      FLB_CONFIG_MAP_TIME, "collector.netdev.scrape_interval", "0",
-     0, FLB_TRUE, offsetof(struct flb_ne, netdev_scrape_interval),
+     0, FLB_FALSE, 0,
      "scrape interval to collect netdev metrics from the node."
     },
 
     {
      FLB_CONFIG_MAP_TIME, "collector.filefd.scrape_interval", "0",
-     0, FLB_TRUE, offsetof(struct flb_ne, filefd_scrape_interval),
+     0, FLB_FALSE, 0,
      "scrape interval to collect filefd metrics from the node."
     },
 
     {
      FLB_CONFIG_MAP_TIME, "collector.textfile.scrape_interval", "0",
-     0, FLB_TRUE, offsetof(struct flb_ne, textfile_scrape_interval),
+     0, FLB_FALSE, 0,
      "scrape interval to collect textfile metrics from the node."
     },
 
     {
      FLB_CONFIG_MAP_TIME, "collector.systemd.scrape_interval", "0",
-     0, FLB_TRUE, offsetof(struct flb_ne, systemd_scrape_interval),
+     0, FLB_FALSE, 0,
      "scrape interval to collect systemd metrics from the node."
     },
 
     {
      FLB_CONFIG_MAP_TIME, "collector.processes.scrape_interval", "0",
-     0, FLB_TRUE, offsetof(struct flb_ne, processes_scrape_interval),
+     0, FLB_FALSE, 0,
      "scrape interval to collect processes metrics from the node."
     },
