+++ plugins/in_node_exporter_metrics/ne_cpu.c
@@ -22,4 +22,30 @@
 #include "ne_cpufreq_linux.c"
 #elif __APPLE__
 #include "ne_cpu_darwin.c"
+
+struct flb_ne_collector cpufreq_collector = {
+    .name = "cpufreq",
+    .cb_init = NULL,
+    .cb_update = NULL,
+    .cb_exit = NULL
+};
+
+#else
+
+#include "ne.h"
+
+struct flb_ne_collector cpu_collector = {
+    .name = "cpu",
+    .cb_init = NULL,
+    .cb_update = NULL,
+    .cb_exit = NULL
+};
+
+struct flb_ne_collector cpufreq_collector = {
+    .name = "cpufreq",
+    .cb_init = NULL,
+    .cb_update = NULL,
+    .cb_exit = NULL
+};
+
 #endif
