+++ plugins/in_node_exporter_metrics/ne_diskstats_darwin.c
@@ -462,19 +462,20 @@ static int diskstats_update(struct flb_ne *ctx)
     return 0;
 }
 
-int ne_diskstats_init(struct flb_ne *ctx)
+static int ne_diskstats_init(struct flb_ne *ctx)
 {
     ne_diskstats_configure(ctx);
     return 0;
 }
 
-int ne_diskstats_update(struct flb_ne *ctx)
+static int ne_diskstats_update(struct flb_input_instance *ins, struct flb_config *config, void *in_context)
 {
+    struct flb_ne *ctx = (struct flb_ne *)in_context;
     diskstats_update(ctx);
     return 0;
 }
 
-int ne_diskstats_exit(struct flb_ne *ctx)
+static int ne_diskstats_exit(struct flb_ne *ctx)
 {
     flb_free(ctx->dt_metrics);
     if (ctx->dt_regex_skip_devices) {
@@ -482,3 +483,11 @@ int ne_diskstats_exit(struct flb_ne *ctx)
     }
     return 0;
 }
+
+
+struct flb_ne_collector diskstats_collector = {
+    .name = "diskstats",
+    .cb_init = ne_diskstats_init,
+    .cb_update = ne_diskstats_update,
+    .cb_exit = ne_diskstats_exit
+};
