+++ plugins/in_node_exporter_metrics/ne_filefd.h
@@ -22,19 +22,6 @@
 
 #include "ne.h"
 
-#ifdef __linux__
-int ne_filefd_init(struct flb_ne *ctx);
-int ne_filefd_update(struct flb_ne *ctx);
-#else
-static int ne_filefd_init(struct flb_ne *ctx)
-{
-    return 0;
-}
-
-static int ne_filefd_update(struct flb_ne *ctx)
-{
-    return 0;
-}
-#endif
+extern struct flb_ne_collector filefd_collector;
 
 #endif
