+++ plugins/in_node_exporter_metrics/ne_filesystem.h
@@ -22,25 +22,6 @@
 
 #include "ne.h"
 
-#ifdef __linux__
-int ne_filesystem_init(struct flb_ne *ctx);
-int ne_filesystem_update(struct flb_ne *ctx);
-int ne_filesystem_exit(struct flb_ne *ctx);
-#else
-static int ne_filesystem_init(struct flb_ne *ctx)
-{
-    return 0;
-}
-
-static int ne_filesystem_update(struct flb_ne *ctx)
-{
-    return 0;
-}
-
-static int ne_filesystem_exit(struct flb_ne *ctx)
-{
-    return 0;
-}
-#endif
+extern struct flb_ne_collector filesystem_collector;
 
 #endif
