+++ plugins/in_node_exporter_metrics/ne_loadavg.c
@@ -21,4 +21,15 @@
 #include "ne_loadavg_linux.c"
 #elif __APPLE__
 #include "ne_loadavg_darwin.c"
+#else
+
+#include "ne.h"
+
+struct flb_ne_collector loadavg_collector = {
+    .name = "loadavg",
+    .cb_init = NULL,
+    .cb_update = NULL,
+    .cb_exit = NULL
+};
+
 #endif
