+++ plugins/in_node_exporter_metrics/ne_loadavg_linux.c
@@ -113,14 +113,23 @@ static int loadavg_update(struct flb_ne *ctx)
     return 0;
 }
 
-int ne_loadavg_init(struct flb_ne *ctx)
+static int ne_loadavg_init(struct flb_ne *ctx)
 {
     ne_loadavg_configure(ctx);
     return 0;
 }
 
-int ne_loadavg_update(struct flb_ne *ctx)
+static int ne_loadavg_update(struct flb_input_instance *ins, struct flb_config *config, void *in_context)
 {
+    struct flb_ne *ctx = (struct flb_ne *)in_context;
+
     loadavg_update(ctx);
     return 0;
 }
+
+struct flb_ne_collector loadavg_collector = {
+    .name = "loadavg",
+    .cb_init = ne_loadavg_init,
+    .cb_update = ne_loadavg_update,
+    .cb_exit = NULL
+};
