+++ plugins/in_node_exporter_metrics/ne_meminfo_linux.c
@@ -262,22 +262,30 @@ static int meminfo_update(struct flb_ne *ctx)
     return 0;
 }
 
-int ne_meminfo_init(struct flb_ne *ctx)
+static int ne_meminfo_init(struct flb_ne *ctx)
 {
     meminfo_configure(ctx);
     return 0;
 }
 
-int ne_meminfo_update(struct flb_ne *ctx)
+static int ne_meminfo_update(struct flb_input_instance *ins, struct flb_config *config, void *in_context)
 {
+    struct flb_ne *ctx = (struct flb_ne *)in_context;
     meminfo_update(ctx);
     return 0;
 }
 
-int ne_meminfo_exit(struct flb_ne *ctx)
+static int ne_meminfo_exit(struct flb_ne *ctx)
 {
     if (ctx->meminfo_ht) {
         flb_hash_table_destroy(ctx->meminfo_ht);
     }
     return 0;
 }
+
+struct flb_ne_collector meminfo_collector = {
+    .name = "meminfo",
+    .cb_init = ne_meminfo_init,
+    .cb_update = ne_meminfo_update,
+    .cb_exit = ne_meminfo_exit
+};
