+++ plugins/in_node_exporter_metrics/ne_netdev.c
@@ -21,4 +21,15 @@
 #include "ne_netdev_linux.c"
 #elif __APPLE__
 #include "ne_netdev_darwin.c"
+#else
+
+#include "ne.h"
+
+struct flb_ne_collector netdev_collector = {
+    .name = "netdev",
+    .cb_init = NULL,
+    .cb_update = NULL,
+    .cb_exit = NULL
+};
+
 #endif
