+++ plugins/in_node_exporter_metrics/ne_processes.h
@@ -22,26 +22,6 @@
 
 #include "ne.h"
 
-#ifdef __linux__
-int ne_processes_init(struct flb_ne *ctx);
-int ne_processes_update(struct flb_ne *ctx);
-int ne_processes_exit(struct flb_ne *ctx);
-#else
-static int ne_processes_init(struct flb_ne *ctx)
-{
-    return 0;
-}
-
-static int ne_processes_update(struct flb_ne *ctx)
-{
-    return 0;
-}
-
-static int ne_processes_exit(struct flb_ne *ctx)
-{
-    return 0;
-}
-
-#endif
+extern struct flb_ne_collector processes_collector;
 
 #endif
