+++ plugins/in_node_exporter_metrics/ne_stat_linux.c
@@ -139,14 +139,23 @@ static int stat_update(struct flb_ne *ctx)
     return 0;
 }
 
-int ne_stat_init(struct flb_ne *ctx)
+static int ne_stat_init(struct flb_ne *ctx)
 {
     stat_configure(ctx);
     return 0;
 }
 
-int ne_stat_update(struct flb_ne *ctx)
+static int ne_stat_update(struct flb_input_instance *ins, struct flb_config *config, void *in_context)
 {
+    struct flb_ne *ctx = (struct flb_ne *)in_context;
+
     stat_update(ctx);
     return 0;
 }
+
+struct flb_ne_collector stat_collector = {
+    .name = "stat",
+    .cb_init = ne_stat_init,
+    .cb_update = ne_stat_update,
+    .cb_exit = NULL
+};
