+++ plugins/in_node_exporter_metrics/ne_systemd.h
@@ -2,7 +2,7 @@
 
 /*  Fluent Bit
  *  ==========
- *  Copyright (C) 2015-2022 The Fluent Bit Authors
+ *  Copyright (C) 2015-2023 The Fluent Bit Authors
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -22,106 +22,6 @@
 
 #include "ne.h"
 
-#define SYSTEMD_UNIT_TYPE_UNDEFINED 0
-#define SYSTEMD_UNIT_TYPE_SERVICE   1
-#define SYSTEMD_UNIT_TYPE_SOCKET    2
-#define SYSTEMD_UNIT_TYPE_MOUNT     3
-#define SYSTEMD_UNIT_TYPE_TIMER     4
+extern struct flb_ne_collector systemd_collector;
 
-struct ne_systemd_unit {
-    char     *name;
-    char     *description;
-    char     *load_state;
-    char     *active_state;
-    char     *sub_state;
-    char     *followed;
-    char     *path;
-    uint32_t  job_id;
-    char     *job_type;
-    char     *object_path;
-
-    /* not part of the unit list result */
-    uint64_t  start_time;
-    int       unit_type;
-    char     *type;
-
-    /* services */
-    uint32_t  restart_count;
-    uint64_t  active_tasks;
-    uint64_t  max_tasks;
-
-    /* sockets */
-    uint32_t  accepted_connections;
-    uint32_t  active_connections;
-    uint32_t  refused_connections;
-
-    /* timers */
-    uint64_t  last_trigger_timestamp;
-};
-
-#ifdef FLB_HAVE_SYSTEMD_SDBUS
-int ne_systemd_init(struct flb_ne *ctx);
-int ne_systemd_update(struct flb_ne *ctx);
-int ne_systemd_exit(struct flb_ne *ctx);
-#else
-static int ne_systemd_init(struct flb_ne *ctx)
-{
-    return 0;
-}
-static int ne_systemd_update(struct flb_ne *ctx)
-{
-    return 0;
-}
-static int ne_systemd_exit(struct flb_ne *ctx)
-{
-    return 0;
-}
-#endif
-
-#define get_system_state(context, output_variable) \
-            get_system_property(context, NULL, "SystemState", \
-                                's', (void *) (output_variable))
-
-#define get_system_version(context, output_variable) \
-            get_system_property(context, NULL, "Version", \
-                                's', (void *) (output_variable))
-
-#define get_service_type(context, unit, output_variable) \
-            get_unit_property(context, unit, NULL, "Type", \
-                              's', (void *) (output_variable))
-
-#define get_service_active_tasks(context, unit, output_variable) \
-            get_unit_property(context, unit, NULL, "TasksCurrent", \
-                              't', (void *) (output_variable))
-
-#define get_service_max_tasks(context, unit, output_variable) \
-            get_unit_property(context, unit, NULL, "TasksMax", \
-                              't', (void *) (output_variable))
-
-#define get_service_restart_count(context, unit, output_variable) \
-            get_unit_property(context, unit, NULL, "NRestarts", \
-                              'u', (void *) (output_variable))
-
-#define get_socket_accepted_connection_count(context, unit, output_variable) \
-            get_unit_property(context, unit, NULL, "NAccepted", \
-                              'u', (void *) (output_variable))
-
-#define get_socket_active_connection_count(context, unit, output_variable) \
-            get_unit_property(context, unit, NULL, "NConnections", \
-                              'u', (void *) (output_variable))
-
-#define get_socket_refused_connection_count(context, unit, output_variable) \
-            get_unit_property(context, unit, NULL, "NRefused", \
-                              'u', (void *) (output_variable))
-
-#define get_timer_last_trigger_timestamp(context, unit, output_variable) \
-            get_unit_property(context, unit, NULL, "LastTriggerUSec", \
-                              't', (void *) (output_variable))
-
-#define get_unit_start_time(context, unit, output_variable) \
-            get_unit_property(context, \
-                              unit, \
-                              "org.freedesktop.systemd1.Unit", \
-                              "ActiveEnterTimestamp", \
-                              't', (void *) (output_variable))
 #endif
