+++ plugins/in_node_exporter_metrics/ne_uname.c
@@ -21,4 +21,15 @@
 #include "ne_uname_linux.c"
 #elif __APPLE__
 #include "ne_uname_darwin.c"
+#else
+
+#include "ne.h"
+
+struct flb_ne_collector uname_collector = {
+    .name = "uname",
+    .cb_init = NULL,
+    .cb_update = NULL,
+    .cb_exit = NULL
+};
+
 #endif
