+++ plugins/in_node_exporter_metrics/ne_uname.h
@@ -22,19 +22,6 @@
 
 #include "ne.h"
 
-#if defined(__linux__) || defined(__APPLE__)
-int ne_uname_init(struct flb_ne *ctx);
-int ne_uname_update(struct flb_ne *ctx);
-#else
-static int ne_uname_init(struct flb_ne *ctx)
-{
-    return 0;
-}
-
-static int ne_uname_update(struct flb_ne *ctx)
-{
-    return 0;
-}
-#endif
+extern struct flb_ne_collector uname_collector;
 
 #endif
