+++ plugins/in_node_exporter_metrics/ne_uname_linux.c
@@ -71,14 +71,23 @@ static int uname_update(struct flb_ne *ctx)
     return ret;
 }
 
-int ne_uname_init(struct flb_ne *ctx)
+static int ne_uname_init(struct flb_ne *ctx)
 {
     uname_configure(ctx);
     return 0;
 }
 
-int ne_uname_update(struct flb_ne *ctx)
+static int ne_uname_update(struct flb_input_instance *ins, struct flb_config *config, void *in_context)
 {
+    struct flb_ne *ctx = (struct flb_ne *)in_context;
+
     uname_update(ctx);
     return 0;
 }
+
+struct flb_ne_collector uname_collector = {
+    .name = "uname",
+    .cb_init = ne_uname_init,
+    .cb_update = ne_uname_update,
+    .cb_exit = NULL
+};
